reformats routes and controllers for login
