+++ models/favorites.js
@@ -31,3 +31,5 @@ class Favorite {
         })
     }
 }
+
+module.exports = Favorite;
