+++ models/methods.js
@@ -28,3 +28,5 @@ class Method {
         })
     }
 }
+
+module.exports = Method;
